rename service folder to services
