+++ .gitignore
@@ -10,6 +10,5 @@ node_modules
 .npm/
 .ssh/
 .viminfo
-local/
 coverage/
 dist/
