+++ src/index.ts
@@ -21,6 +21,6 @@
 
 export { default as Client } from "./client";
 export { default as Config } from "./config";
-export * from "./service/";
+export * from "./services/";
 export { default as utils } from "./utils";
 export { default as HttpURLConnectionClient } from "./httpClient/httpURLConnectionClient";
