+++ src/services/binLookup.ts
