+++ src/services/resource/modification/capture.ts
